requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pyyaml>=6.0
mirascope[openai]>=1.22.6
rich>=13.7
tiktoken>=0.6
oxen
mirascope 
ptyprocess
pydantic